rename template directives and attempting to ignore them in codeclimate
