+++ projects/ng-sq-ui/src/lib/form-elements/checkbox/checkbox.template.directive.ts
