+++ projects/ng-sq-ui/src/lib/form-elements/typeahead/typeahead.template.directive.ts
